+++ host/lib/usrp/multi_crimson_tng.cpp
@@ -488,7 +488,7 @@ static bool printed_get_time_now;
 // Get the current time on Crimson
 time_spec_t multi_crimson_tng::get_time_now(size_t mboard){
 	crimson_tng_impl::sptr dev = boost::static_pointer_cast<crimson_tng_impl>( _dev );
-	double diff = NULL == dev.get() ? 0 : dev.get()->get_time_diff();
+	double diff = NULL == dev.get() ? 0 : dev.get()->time_diff_get();
 	if ( ! printed_get_time_now ) {
 		if ( NULL != dev.get() ) {
 			std::cout << "time diff is " << diff << std::endl;
